requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==19.11.2
Automat==24.8.1
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
charset-normalizer==3.1.0
click==8.1.7
comm==0.1.4
constantly==23.10.4
contourpy==1.1.0
cryptography==43.0.0
cycler==0.11.0
debugpy==1.6.8
decorator==5.1.1
docutils @ file:///private/tmp/docutils-20231211-36728-9mhbst/docutils-0.20.1
executing==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiteconnect==5.0.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
Pygments==2.15.1
pyOpenSSL==24.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scikit-learn-extra==0.3.0
scipy==1.11.1
service-identity==24.1.0
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.15.0
yarl==1.9.2
zope.interface==7.0.3
